+++ src/systemmonitor.h
@@ -9,6 +9,8 @@ class SystemMonitor : public QObject {
   Q_OBJECT
   Q_PROPERTY(double cpuUsage READ cpuUsage NOTIFY cpuUsageChanged)
   Q_PROPERTY(double memoryUsage READ memoryUsage NOTIFY memoryUsageChanged)
+  Q_PROPERTY(QString totalMemory READ totalMemory NOTIFY totalMemoryChanged)
+  Q_PROPERTY(QString usedMemory READ usedMemory NOTIFY usedMemoryChanged)
   Q_PROPERTY(QString networkStats READ networkStats NOTIFY networkStatsChanged)
   Q_PROPERTY(double systemLoad READ systemLoad NOTIFY systemLoadChanged)
   Q_PROPERTY(bool active READ active WRITE setActive NOTIFY activeChanged)
@@ -19,6 +21,8 @@ public:
 
   double cpuUsage() const { return m_cpuUsage; }
   double memoryUsage() const { return m_memoryUsage; }
+  QString totalMemory() const { return m_totalMemory; }
+  QString usedMemory() const { return m_usedMemory; }
   QString networkStats() const { return m_networkStats; }
   double systemLoad() const { return m_systemLoad; }
   bool active() const { return m_active; }
@@ -28,6 +32,8 @@ public:
 signals:
   void cpuUsageChanged();
   void memoryUsageChanged();
+  void totalMemoryChanged();
+  void usedMemoryChanged();
   void networkStatsChanged();
   void systemLoadChanged();
   void activeChanged();
@@ -45,6 +51,8 @@ private:
   QTimer *m_timer;
   double m_cpuUsage;
   double m_memoryUsage;
+  QString m_totalMemory;
+  QString m_usedMemory;
   QString m_networkStats;
   double m_systemLoad;
   bool m_active;